Update privacy.pug
Minimum age to create an account 13 > 16 as in accordance with Art. 8 GDPR. The age of consent can be lowered and does differ throughout the Union as countries can lower the age. But as stipulated by the GDPR 16 years should be the minimum, unless you want to go ahead and list the minimum age for each country.